+++ vcd/config.go
@@ -25,6 +25,16 @@ type Config struct {
 	Href            string
 	MaxRetryTimeout int
 	InsecureFlag    bool
+
+	// UseSamlAdfs specifies if SAML auth is used for authenticating vCD instead of local login.
+	// The following conditions must be met so that authentication SAML authentication works:
+	// * SAML IdP (Identity Provider) is Active Directory Federation Service (ADFS)
+	// * Authentication endpoint "/adfs/services/trust/13/usernamemixed" must be enabled on ADFS
+	// server
+	UseSamlAdfs bool
+	// CustomAdfsRptId allows to set custom Relaying Party Trust identifier. By default vCD Entity
+	// ID is used as Relaying Party Trust identifier.
+	CustomAdfsRptId string
 }
 
 type VCDClient struct {
@@ -390,7 +400,8 @@ func (c *Config) Client() (*VCDClient, error) {
 
 	vcdClient := &VCDClient{
 		VCDClient: govcd.NewVCDClient(*authUrl, c.InsecureFlag,
-			govcd.WithMaxRetryTimeout(c.MaxRetryTimeout)),
+			govcd.WithMaxRetryTimeout(c.MaxRetryTimeout),
+			govcd.WithSamlAdfs(c.UseSamlAdfs, c.CustomAdfsRptId)),
 		SysOrg:          c.SysOrg,
 		Org:             c.Org,
 		Vdc:             c.Vdc,
@@ -404,11 +415,6 @@ func (c *Config) Client() (*VCDClient, error) {
 	cachedVCDClients.Lock()
 	cachedVCDClients.conMap[checksum] = cachedConnection{initTime: time.Now(), connection: vcdClient}
 	cachedVCDClients.Unlock()
-
-	// Throw a deprecated warning for vCD versions <= 9.1
-	if vcdClient.Client.APIVCDMaxVersionIs("<= 30") {
-		_, _ = fmt.Fprintln(getTerraformStdout(), "WARNING: Support for vCD versions older than 9.5 is deprecated")
-	}
 
 	return vcdClient, nil
 }